requirements.txt
discord.pyz
python-dotenv